+++ src/main/kotlin/io/github/enyason/client/polling/PollingStrategy.kt
@@ -1,6 +1,6 @@
 package io.github.enyason.client.polling
 
-import io.github.enyason.client.Task
+import io.github.enyason.client.task.Task
 
 /**
  * An interface defining a strategy for polling the status and results of a [Task].
